Conflicts:
	src/main/java/nl/webedu/dao/UserStoryDAO.java
	src/main/java/nl/webedu/resources/UserStoryResource.java

+++ src/main/java/nl/webedu/services/ProjectService.java
@@ -1,16 +1,9 @@
 package nl.webedu.services;
 
-import com.google.common.base.Optional;
-import io.dropwizard.auth.Auth;
 import java.util.ArrayList;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.validation.Valid;
-import javax.ws.rs.QueryParam;
 import nl.webedu.dao.ProjectDAO;
 import nl.webedu.models.EmployeeModel;
 import nl.webedu.models.ProjectModel;
-import nl.webedu.resources.ProjectResource;
 
 public class ProjectService {
     private ProjectDAO projectDAO;
@@ -26,20 +19,32 @@ public class ProjectService {
         return this.projectDAO.project_list_employee(employeeId);
     }
     
-    public boolean createProject(ProjectModel projectModel){
-        return projectDAO.addProject(projectModel);
+    public boolean createProject(ProjectModel projectModel, EmployeeModel loggedUser){
+        if(loggedUser.getEmployeeRole().equals("administration")){
+            return projectDAO.addProject(projectModel);
+        }else{return false;}
+        
     }
         
-    public boolean update(ProjectModel projectModel) throws Exception{
-         return projectDAO.modifyProject(projectModel);
+    public boolean update(ProjectModel projectModel, EmployeeModel loggedUser) throws Exception{
+        if(loggedUser.getEmployeeRole().equals("administration")){
+           return projectDAO.modifyProject(projectModel); 
+        }else{return false;}
+         
     }
 
-    public boolean delete(ProjectModel projectModel){
-        return projectDAO.removeProject(projectModel);
+    public boolean delete(ProjectModel projectModel, EmployeeModel loggedUser){
+        if(loggedUser.getEmployeeRole().equals("administration")){
+            return projectDAO.removeProject(projectModel);
+        }else{return false;}
+        
     }
 
-    public boolean unDeleteByUrl(int projectId){
-        return this.projectDAO.unRemoveProject(projectId);
+    public boolean unDeleteByUrl(int projectId, EmployeeModel loggedUser){
+        if(loggedUser.getEmployeeRole().equals("administration")){
+            return this.projectDAO.unRemoveProject(projectId);
+        }else{return false;}
+        
     }
